Fix tests being treated as "pending" if they break before assert()
If a test errors before it gets to an assertion, it treats the test as
pending rather than being broken due the function being exited out.

We're taking a different approach to error handling now, where if
there's something wrong when the function is called, you get a normal
error traceback.